remove trends search duplicates
