--- src/components/Trends/Explore/TrendsExploreSearch.module.scss
@@ -1,25 +1,0 @@
-@import '~@santiment-network/ui/mixins';
-
-.wrapper {
-  @include responsive('tablet', 'laptop', 'desktop') {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-
-  @include responsive('phone-xs', 'phone') {
-    width: 100%;
-  }
-}
-
-.input {
-  width: 100%;
-
-  @include responsive('phone-xs', 'phone') {
-    margin-right: 0;
-  }
-
-  & input {
-    height: 40px;
-  }
-}
