+++ src/components/Trends/Search/Suggestions.module.scss
@@ -1,11 +1,5 @@
 @import '~@santiment-network/ui/mixins';
 
-.TrendsSearch {
-  display: flex;
-  flex-direction: column;
-  margin-top: 26px;
-}
-
 .examples {
   margin-top: 8px;
   display: flex;
